refactor: Name객체 제거 및 원시값으로 변환
