--- src/main/java/Name.java
@@ -1,15 +1,0 @@
-public class Name {
-
-    String name;
-
-    public Name(String name) {
-        validateNameLength(name);
-        this.name = name;
-    }
-
-    private  void validateNameLength(String name) {
-        if (name.length() > 5) {
-            throw new IllegalArgumentException("[ERROR] 사람 이름은 최대 5글자 입니다.");
-        }
-    }
-}
