+++ src/test/java/PlayerTest.java
@@ -1,3 +1,4 @@
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import org.junit.jupiter.api.DisplayName;
@@ -5,10 +6,21 @@ import org.junit.jupiter.api.Test;
 
 public class PlayerTest {
 
-    @DisplayName("Player 객체 생성")
+    @DisplayName("5글자이면 player 객체가 생성")
     @Test
-    void createPlayer() {
-        Name roy = new Name("roy");
-        assertDoesNotThrow(() -> new Player(roy));
+    void noError() {
+        String inputName = "abcde";
+        assertDoesNotThrow(() -> new Player(inputName));
+
     }
+
+    @DisplayName("6글자 이상 입력될 경우 예외 발생")
+    @Test
+    void errorAboutNameLimit() {
+        String inputName = "abcdef";
+        assertThatThrownBy(() ->
+                new Player(inputName)).isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("[ERROR] 사람 이름은 최대 5글자 입니다.");
+    }
+
 }
